+++ data-structure-algorithm/src/main/java/com/algorithm/dynamicprogramming/topdown/HouseThief.java
@@ -4,13 +4,13 @@ import com.algorithm.IHouseThief;
 
 public class HouseThief implements IHouseThief {
 	
-	public int getMaxValue(int[] memory, int[] houses, int currentIndex) {
+	public long getMaxValue(long[] memory, int[] houses, int currentIndex) {
 		if (currentIndex >= houses.length) {
-			return 0;
+			return 0L;
 		}
 		if (memory[currentIndex] == 0) {
-			int choice1 = houses[currentIndex] + getMaxValue(memory, houses, currentIndex + 2);
-			int choice2 = getMaxValue(memory, houses, currentIndex + 1);
+			long choice1 = houses[currentIndex] + getMaxValue(memory, houses, currentIndex + 2);
+			long choice2 = getMaxValue(memory, houses, currentIndex + 1);
 
 			memory[currentIndex] = (choice1 > choice2) ? choice1 : choice2;
 		}
@@ -19,7 +19,7 @@ public class HouseThief implements IHouseThief {
 
 	@Override
 	public long maxStealth(int[] wealth) {
-		int[] memory = new int[wealth.length];
+		long[] memory = new long[wealth.length];
 		
 		return getMaxValue(memory, wealth, 0);
 	}
